Fix build issues and compiler warnings with MSVC 2013

+++ include/valijson/validation_visitor.hpp
@@ -651,7 +651,7 @@ public:
                 }
                 return false;
             }
-            d = i;
+            d = static_cast<double>(i);
         } else {
             return true;
         }
